фильтры разбиты на подкомпоненты
