+++ app/src/App.js
@@ -1,7 +1,7 @@
 
 import "./css/style.css";
 import {connect} from "react-redux";
-import Filters from "./components/Filters";
+import Filters from "./components/Filters/Filters";
 import Table from "./components/Table";
 import Pagination from "./components/Pagination";
 import {ajaxGetDataForFilters} from "./actions/ajax";
@@ -11,7 +11,7 @@ import {useEffect} from "react";
 const App = (props) => {
   useEffect(() => {
     props.onAjaxGetDataForFilters();
-  })
+  },[])
   return (
       <section className="blog-area section-padding-100-0">
           <div className="container">
