+++ app/src/actions/table.js
@@ -1,0 +1,2 @@
+
+export const setData = (data) => ({ type: 'SET_DATA', data});
