+++ app/src/reducers/index.js
@@ -1,6 +1,8 @@
 import {combineReducers} from "redux";
 import filters from "./filters";
+import table from "./table";
 
 export default combineReducers({
     filters,
+    table
 })
